Move assign_roles to Game
Avoids a dependency loop
