+++ src/Avalon/Game.py
@@ -2,7 +2,24 @@ import random
 
 import discord
 
-from src.Avalon.Roles.Role import assign_role
+from src.Avalon.Roles.LoyalServant import LoyalServant
+from src.Avalon.Roles.Merlin import Merlin
+from src.Avalon.Roles.Minion import Minion
+from src.Avalon.Roles.Morgana import Morgana
+from src.Avalon.Roles.Percival import Percival
+
+
+def assign_role(player, role_name):
+    if role_name is "Loyal Servant":
+        return LoyalServant(player)
+    if role_name is "Morgana":
+        return Morgana(player)
+    if role_name is "Merlin":
+        return Merlin(player)
+    if role_name is "Percival":
+        return Percival(player)
+    if role_name is "Loathsome Minion of Mordred":
+        return Minion(player)
 
 
 class Game:
